#12 Restructure project
