+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Dashboard from "./Dashboard";
+import Dashboard from "./Components/Dashboard";
 
 const App = () => {
   return (
